@@ -1,7 +1,7 @@
 [project]
 name = "polymarket-apis"
-version = "0.1.1"
-description = "WIP - Unified Polymarket apis - clob, gamma, data, web3"
+version = "0.1.2"
+description = "WIP - Unified Polymarket apis - clob, gamma, data, web3, websockets"
 authors = [
     { name = "Razvan Gheorghe", email = "razvan@gheorghe.me" }
 ]
@@ -14,9 +14,10 @@ dependencies = [
     "gql>=3.5.2",
     "lomond>=0.3.3",
     "wsaccel>=0.6.7",
+
 ]
 readme = "README.md"
-requires-python = ">= 3.8"
+requires-python = ">=3.10"
 
 [build-system]
 requires = ["hatchling"]
